fix compiler warnings when building with /Wall in Visual Studio

+++ modules/c++/sys/include/sys/ConditionVarWin32.h
@@ -57,7 +57,7 @@ private:
 
 private:
     // # of waiting threads
-    size_t           mNumWaiters;
+    LONG           mNumWaiters;
     CRITICAL_SECTION mNumWaitersCS;
 
     // Semaphore used to queue up threads waiting for the condition to
